+++ .eslintrc
@@ -6,15 +6,11 @@
     "@mskelton",
     "@mskelton/eslint-config/typescript",
     "plugin:@typescript-eslint/strict",
-    "plugin:promise/recommended",
-    "plugin:jest/all"
+    "plugin:promise/recommended"
   ],
   "rules": {
     "no-return-await": "warn",
     "prefer-template": "warn",
-    "curly": ["warn", "all"],
-    "jest/unbound-method": "off",
-    "jest/prefer-expect-assertions": "off",
-    "jest/prefer-spy-on": "off"
+    "curly": ["warn", "all"]
   }
 }
