+++ .vscodeignore
@@ -11,7 +11,6 @@ locales/bundle.l10n.jsonc
 .prettierignore
 .gitignore
 
-jest.config.js
 tsconfig.json
 
 out/**
