+++ package.json
@@ -28,7 +28,7 @@
     "lint": "eslint src --ext ts",
     "format:check": "prettier --check .",
     "format": "prettier --write .",
-    "test": "NODE_OPTIONS=--experimental-vm-modules npx jest",
+    "test": "vitest",
     "test:watch": "npm run test -- --watch"
   },
   "icon": "assets/icon.png",
@@ -56,13 +56,11 @@
     "@typescript-eslint/parser": "^5.48.2",
     "esbuild": "^0.17.3",
     "eslint": "^8.32.0",
-    "eslint-plugin-jest": "^27.2.1",
     "eslint-plugin-promise": "^6.1.1",
-    "jest": "^29.3.1",
     "prettier": "^2.8.3",
     "semantic-release-vsce": "^5.5.5",
-    "ts-jest": "^29.0.5",
-    "typescript": "^4.9.4"
+    "typescript": "^4.9.4",
+    "vitest": "^0.28.4"
   },
   "dependencies": {
     "semver": "^7.3.8"
